requirements.txt
bleak==0.22.1
toio.py @ git+https://github.com/toio/toio.py
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
